💡 Fix timestamps throwing off packwiz refresh
These are artificial changes to make sure that git will pick up a change and propogate it to all other users' repos.  Currently, the contents of these files are identical on the dev and deployed machine, but the timestamps are different, causing packwiz to generate different hash values when it re-indexes on a refresh, which it does automatically.